# receiver/openapi.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: Star Wars Quiz Game API
  contact:
    email: anandanmol150@gmail.com
    
paths:
  /questions:
    get:
      summary: gets 5 questions for the quiz to start
      operationId: app.get_questions
      responses:
        '200':
          description: Sucessfully returned 5 questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: string
  /score:
    post:
      summary: saves score
      operationId: app.store_score
      description: Saves a score from our quiz game
      responses:
        '201':
          description: score saved
        '400':
          description: 'invalid input, object invalid'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Score'
        description: Score to save
  /leaderboard:
    get:
      summary: gets top 10 scores for the leaderboard
      operationId: app.get_top_scores
      responses:
        '200':
          description: Sucessfully returned 10 scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Score'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: string

components:
  schemas:
    Question:
      required:
      - question
      - correct_answer
      - options
      properties:
        question:
          type: string
          example: "What's your name?"
        correct_answer:
          example: "Anmol"
        options:
          type: array
    Score:
      type: object
      required:
        - name
        - score
      properties:
        name:
          type: string
          example: 'Anmol'
        score:
          type: integer
          example: 3